pyproject.toml
[tool.poetry]
name = "sbbot"
version = "0.1.0"
description = ""
authors = ["Annekin Meyburgh <annekin.meyburgh@takealot.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
loguru = "^0.6.0"
PyYAML = "^6.0"
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
